Refactor event bus to reduce fragmentation
Merged `plugin/framework/events.py` into `plugin/framework/event_bus.py`.
Updated imports across the test suite to use the unified module.
