+++ plugin/framework/event_bus.py
@@ -19,6 +19,7 @@
 import logging
 import weakref
 
+from plugin.framework.service_base import ServiceBase
 log = logging.getLogger("writeragent.events")
 
 
@@ -121,3 +122,17 @@ def get_event_bus():
     return sys._localwriter_event_bus
 
 global_event_bus = get_event_bus()
+
+
+class EventBusService(ServiceBase, EventBus):
+    """Singleton event bus exposed as a service.
+
+    Inherits from both ServiceBase (for registry) and EventBus (for
+    pub/sub). Modules access it as ``services.events``.
+    """
+
+    name = "events"
+
+    def __init__(self):
+        ServiceBase.__init__(self)
+        self._subscribers = global_event_bus._subscribers
