+++ plugin/tests/uno/test_writer_search.py
@@ -68,7 +68,7 @@ class MockWriterIndexService:
 class MockServices:
     def __init__(self, doc):
         from plugin.framework.document import DocumentService
-        from plugin.framework.events import EventBus
+        from plugin.framework.event_bus import EventBus
         self.events = EventBus()
         # DocumentService does not take constructor arguments; it uses the
         # active UNO context when needed.
